+++ src/lib.rs
@@ -113,5 +113,94 @@ impl MidiStreamParser {
     }
 }
 
+/// A MIDI SysEx message with internal buffer.
+#[derive(Debug)]
+pub struct MidiSysexMessage<const CAPACITY: usize> {
+    /// Buffer containing the message.
+    buffer: [u8; CAPACITY],
+
+    /// Length of the message in the buffer.
+    length: usize,
+
+    /// Flag for running message.
+    running: bool,
+}
+
+impl<const CAPACITY: usize> Default for MidiSysexMessage<CAPACITY> {
+    /// Returns a new parser with default values.
+    fn default() -> Self {
+        Self {
+            buffer: [0; CAPACITY],
+            length: 0,
+            running: false,
+        }
+    }
+}
+
+impl<const CAPACITY: usize> MidiSysexMessage<CAPACITY> {
+    /// Returns a new SysEx message.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Appends a byte to the buffer and returns an option
+    /// with a `Some(&[u8])` when the message is complete.
+    pub fn append(&mut self, byte: u8) -> Option<&[u8]> {
+        match byte {
+            0xF0 => {
+                // Start of message.
+                self.buffer[0] = byte;
+                self.length = 1;
+                self.running = true;
+                None
+            }
+            0xF7 if self.running && self.length < CAPACITY => {
+                // End of message.
+                self.buffer[self.length] = byte;
+                self.length += 1;
+                self.running = false;
+                Some(&self.buffer[..self.length])
+            }
+            0x00..=0x7F if self.running && self.length < CAPACITY => {
+                // Data byte.
+                self.buffer[self.length] = byte;
+                self.length += 1;
+                None
+            }
+            _ => None,
+        }
+    }
+
+    /// Clears the data in the buffer.
+    pub fn clear(&mut self) {
+        *self = Self::default();
+    }
+
+    /// Returns the capacity of the buffer.
+    pub fn capacity(&self) -> usize {
+        CAPACITY
+    }
+
+    /// Returns the length of the data in the buffer.
+    pub fn len(&self) -> usize {
+        self.length
+    }
+
+    /// Returns the free capacity of the buffer.
+    pub fn free_capacity(&self) -> usize {
+        CAPACITY - self.len()
+    }
+
+    /// Returns if the buffer is empty.
+    pub fn is_empty(&self) -> bool {
+        self.length == 0
+    }
+
+    /// Returns if the buffer is full.
+    pub fn is_full(&self) -> bool {
+        self.length == CAPACITY
+    }
+}
+
 #[cfg(test)]
 mod tests;